separate logger implementation for each platfor to diffrent file
