+++ esphome/components/logger/logger.cpp
@@ -1,28 +1,6 @@
 #include "logger.h"
 #include <cinttypes>
 
-#ifdef USE_ESP_IDF
-#include <driver/uart.h>
-
-#if defined(USE_ESP32_VARIANT_ESP32C3) || defined(USE_ESP32_VARIANT_ESP32C6) || defined(USE_ESP32_VARIANT_ESP32S3) || \
-    defined(USE_ESP32_VARIANT_ESP32H2)
-#include <driver/usb_serial_jtag.h>
-#include <esp_vfs_dev.h>
-#include <esp_vfs_usb_serial_jtag.h>
-#endif
-
-#include "freertos/FreeRTOS.h"
-#include "esp_idf_version.h"
-
-#include <cstdint>
-#include <cstdio>
-#include <fcntl.h>
-
-#endif  // USE_ESP_IDF
-
-#if defined(USE_ESP32_FRAMEWORK_ARDUINO) || defined(USE_ESP_IDF)
-#include <esp_log.h>
-#endif  // USE_ESP32_FRAMEWORK_ARDUINO || USE_ESP_IDF
 #include "esphome/core/hal.h"
 #include "esphome/core/log.h"
 
@@ -106,58 +84,6 @@ void Logger::log_vprintf_(int level, const char *tag, int line, const __FlashStr
 }
 #endif
 
-#ifdef USE_ESP_IDF
-void Logger::init_uart_() {
-  uart_config_t uart_config{};
-  uart_config.baud_rate = (int) baud_rate_;
-  uart_config.data_bits = UART_DATA_8_BITS;
-  uart_config.parity = UART_PARITY_DISABLE;
-  uart_config.stop_bits = UART_STOP_BITS_1;
-  uart_config.flow_ctrl = UART_HW_FLOWCTRL_DISABLE;
-#if ESP_IDF_VERSION >= ESP_IDF_VERSION_VAL(5, 0, 0)
-  uart_config.source_clk = UART_SCLK_DEFAULT;
-#endif
-  uart_param_config(this->uart_num_, &uart_config);
-  const int uart_buffer_size = tx_buffer_size_;
-  // Install UART driver using an event queue here
-  uart_driver_install(this->uart_num_, uart_buffer_size, uart_buffer_size, 10, nullptr, 0);
-}
-
-#if defined(USE_ESP32_VARIANT_ESP32S2) || defined(USE_ESP32_VARIANT_ESP32S3)
-void Logger::init_usb_cdc_() {}
-#endif
-
-#if defined(USE_ESP32_VARIANT_ESP32C3) || defined(USE_ESP32_VARIANT_ESP32C6) || defined(USE_ESP32_VARIANT_ESP32S3) || \
-    defined(USE_ESP32_VARIANT_ESP32H2)
-void Logger::init_usb_serial_jtag_() {
-  setvbuf(stdin, NULL, _IONBF, 0);  // Disable buffering on stdin
-
-  // Minicom, screen, idf_monitor send CR when ENTER key is pressed
-  esp_vfs_dev_usb_serial_jtag_set_rx_line_endings(ESP_LINE_ENDINGS_CR);
-  // Move the caret to the beginning of the next line on '\n'
-  esp_vfs_dev_usb_serial_jtag_set_tx_line_endings(ESP_LINE_ENDINGS_CRLF);
-
-  // Enable non-blocking mode on stdin and stdout
-  fcntl(fileno(stdout), F_SETFL, 0);
-  fcntl(fileno(stdin), F_SETFL, 0);
-
-  usb_serial_jtag_driver_config_t usb_serial_jtag_config{};
-  usb_serial_jtag_config.rx_buffer_size = 512;
-  usb_serial_jtag_config.tx_buffer_size = 512;
-
-  esp_err_t ret = ESP_OK;
-  // Install USB-SERIAL-JTAG driver for interrupt-driven reads and writes
-  ret = usb_serial_jtag_driver_install(&usb_serial_jtag_config);
-  if (ret != ESP_OK) {
-    return;
-  }
-
-  // Tell vfs to use usb-serial-jtag driver
-  esp_vfs_usb_serial_jtag_use_driver();
-}
-#endif
-#endif
-
 int HOT Logger::level_for(const char *tag) {
   // Uses std::vector<> for low memory footprint, though the vector
   // could be sorted to minimize lookup times. This feature isn't used that
@@ -169,6 +95,7 @@ int HOT Logger::level_for(const char *tag) {
   }
   return ESPHOME_LOG_LEVEL;
 }
+
 void HOT Logger::log_message_(int level, const char *tag, int offset) {
   // remove trailing newline
   if (this->tx_buffer_[this->tx_buffer_at_ - 1] == '\n') {
@@ -223,179 +150,6 @@ Logger::Logger(uint32_t baud_rate, size_t tx_buffer_size) : baud_rate_(baud_rate
   this->tx_buffer_ = new char[this->tx_buffer_size_ + 1];  // NOLINT
 }
 
-#ifndef USE_LIBRETINY
-void Logger::pre_setup() {
-  if (this->baud_rate_ > 0) {
-#ifdef USE_ARDUINO
-    switch (this->uart_) {
-      case UART_SELECTION_UART0:
-#ifdef USE_ESP8266
-      case UART_SELECTION_UART0_SWAP:
-#endif
-#ifdef USE_RP2040
-        this->hw_serial_ = &Serial1;
-        Serial1.begin(this->baud_rate_);
-#else
-#if ARDUINO_USB_CDC_ON_BOOT
-        this->hw_serial_ = &Serial0;
-        Serial0.begin(this->baud_rate_);
-#else
-        this->hw_serial_ = &Serial;
-        Serial.begin(this->baud_rate_);
-#endif
-#endif
-#ifdef USE_ESP8266
-        if (this->uart_ == UART_SELECTION_UART0_SWAP) {
-          Serial.swap();
-        }
-        Serial.setDebugOutput(ESPHOME_LOG_LEVEL >= ESPHOME_LOG_LEVEL_VERBOSE);
-#endif
-        break;
-      case UART_SELECTION_UART1:
-#ifdef USE_RP2040
-        this->hw_serial_ = &Serial2;
-        Serial2.begin(this->baud_rate_);
-#else
-        this->hw_serial_ = &Serial1;
-        Serial1.begin(this->baud_rate_);
-#endif
-#ifdef USE_ESP8266
-        Serial1.setDebugOutput(ESPHOME_LOG_LEVEL >= ESPHOME_LOG_LEVEL_VERBOSE);
-#endif
-        break;
-#if defined(USE_ESP32) && !defined(USE_ESP32_VARIANT_ESP32C3) && !defined(USE_ESP32_VARIANT_ESP32C6) && \
-    !defined(USE_ESP32_VARIANT_ESP32S2) && !defined(USE_ESP32_VARIANT_ESP32S3)
-      case UART_SELECTION_UART2:
-        this->hw_serial_ = &Serial2;
-        Serial2.begin(this->baud_rate_);
-        break;
-#endif
-#if defined(USE_ESP32) && \
-    (defined(USE_ESP32_VARIANT_ESP32S2) || defined(USE_ESP32_VARIANT_ESP32S3) || defined(USE_ESP32_VARIANT_ESP32C3))
-#if defined(USE_ESP32_VARIANT_ESP32S2) || defined(USE_ESP32_VARIANT_ESP32S3) || defined(USE_ESP32_VARIANT_ESP32C3)
-      case UART_SELECTION_USB_CDC:
-#endif  // USE_ESP32_VARIANT_ESP32S2 || USE_ESP32_VARIANT_ESP32S3 || USE_ESP32_VARIANT_ESP32C3
-#if defined(USE_ESP32_VARIANT_ESP32C3) || defined(USE_ESP32_VARIANT_ESP32S3)
-      case UART_SELECTION_USB_SERIAL_JTAG:
-#endif  // USE_ESP32_VARIANT_ESP32C3 || USE_ESP32_VARIANT_ESP32S3
-#if defined(USE_ESP32_VARIANT_ESP32S2) || defined(USE_ESP32_VARIANT_ESP32S3) || defined(USE_ESP32_VARIANT_ESP32C3)
-#if ARDUINO_USB_CDC_ON_BOOT
-        this->hw_serial_ = &Serial;
-        Serial.setTxTimeoutMs(0);  // workaround for 2.0.9 crash when there's no data connection
-        Serial.begin(this->baud_rate_);
-#else
-        this->hw_serial_ = &Serial;
-        Serial.begin(this->baud_rate_);
-#endif  // ARDUINO_USB_CDC_ON_BOOT
-#endif  // USE_ESP32_VARIANT_ESP32S2 || USE_ESP32_VARIANT_ESP32S3 || USE_ESP32_VARIANT_ESP32C3
-        break;
-#endif  // USE_ESP32 && (USE_ESP32_VARIANT_ESP32S2 || USE_ESP32_VARIANT_ESP32S3 || USE_ESP32_VARIANT_ESP32C3)
-#ifdef USE_RP2040
-      case UART_SELECTION_USB_CDC:
-        this->hw_serial_ = &Serial;
-        Serial.begin(this->baud_rate_);
-        break;
-#endif  // USE_RP2040
-    }
-#endif  // USE_ARDUINO
-#ifdef USE_ESP_IDF
-    this->uart_num_ = UART_NUM_0;
-    switch (this->uart_) {
-      case UART_SELECTION_UART0:
-        this->uart_num_ = UART_NUM_0;
-        break;
-      case UART_SELECTION_UART1:
-        this->uart_num_ = UART_NUM_1;
-        break;
-#if !defined(USE_ESP32_VARIANT_ESP32C3) && !defined(USE_ESP32_VARIANT_ESP32C6) && \
-    !defined(USE_ESP32_VARIANT_ESP32S2) && !defined(USE_ESP32_VARIANT_ESP32S3) && !defined(USE_ESP32_VARIANT_ESP32H2)
-      case UART_SELECTION_UART2:
-        this->uart_num_ = UART_NUM_2;
-        break;
-#endif  // !USE_ESP32_VARIANT_ESP32C3 && !USE_ESP32_VARIANT_ESP32S2 && !USE_ESP32_VARIANT_ESP32S3 &&
-        // !USE_ESP32_VARIANT_ESP32H2
-#if defined(USE_ESP32_VARIANT_ESP32S2) || defined(USE_ESP32_VARIANT_ESP32S3)
-      case UART_SELECTION_USB_CDC:
-        this->uart_num_ = -1;
-        this->init_usb_cdc_();
-        break;
-#endif  // USE_ESP32_VARIANT_ESP32S2 || USE_ESP32_VARIANT_ESP32S3
-#if defined(USE_ESP32_VARIANT_ESP32C3) || defined(USE_ESP32_VARIANT_ESP32C6) || defined(USE_ESP32_VARIANT_ESP32S3) || \
-    defined(USE_ESP32_VARIANT_ESP32H2)
-      case UART_SELECTION_USB_SERIAL_JTAG:
-        this->uart_num_ = -1;
-        this->init_usb_serial_jtag_();
-        break;
-#endif  // USE_ESP32_VARIANT_ESP32C3 || USE_ESP32_VARIANT_ESP32C6 || USE_ESP32_VARIANT_ESP32S3 ||
-        // USE_ESP32_VARIANT_ESP32H2
-    }
-    if (this->uart_num_ >= 0) {
-      this->init_uart_();
-    }
-#endif  // USE_ESP_IDF
-  }
-#ifdef USE_ESP8266
-  else {
-    uart_set_debug(UART_NO);
-  }
-#endif  // USE_ESP8266
-
-  global_logger = this;
-#if defined(USE_ESP_IDF) || defined(USE_ESP32_FRAMEWORK_ARDUINO)
-  esp_log_set_vprintf(esp_idf_log_vprintf_);
-  if (ESPHOME_LOG_LEVEL >= ESPHOME_LOG_LEVEL_VERBOSE) {
-    esp_log_level_set("*", ESP_LOG_VERBOSE);
-  }
-#endif  // USE_ESP_IDF || USE_ESP32_FRAMEWORK_ARDUINO
-
-  ESP_LOGI(TAG, "Log initialized");
-}
-#else  // USE_LIBRETINY
-void Logger::pre_setup() {
-  if (this->baud_rate_ > 0) {
-    switch (this->uart_) {
-#if LT_HW_UART0
-      case UART_SELECTION_UART0:
-        this->hw_serial_ = &Serial0;
-        Serial0.begin(this->baud_rate_);
-        break;
-#endif
-#if LT_HW_UART1
-      case UART_SELECTION_UART1:
-        this->hw_serial_ = &Serial1;
-        Serial1.begin(this->baud_rate_);
-        break;
-#endif
-#if LT_HW_UART2
-      case UART_SELECTION_UART2:
-        this->hw_serial_ = &Serial2;
-        Serial2.begin(this->baud_rate_);
-        break;
-#endif
-      default:
-        this->hw_serial_ = &Serial;
-        Serial.begin(this->baud_rate_);
-        if (this->uart_ != UART_SELECTION_DEFAULT) {
-          ESP_LOGW(TAG, "  The chosen logger UART port is not available on this board."
-                        "The default port was used instead.");
-        }
-        break;
-    }
-
-    // change lt_log() port to match default Serial
-    if (this->uart_ == UART_SELECTION_DEFAULT) {
-      this->uart_ = (UARTSelection) (LT_UART_DEFAULT_SERIAL + 1);
-      lt_log_set_port(LT_UART_DEFAULT_SERIAL);
-    } else {
-      lt_log_set_port(this->uart_ - 1);
-    }
-  }
-
-  global_logger = this;
-  ESP_LOGI(TAG, "Log initialized");
-}
-#endif  // USE_LIBRETINY
-
 void Logger::set_baud_rate(uint32_t baud_rate) { this->baud_rate_ = baud_rate; }
 void Logger::set_log_level(const std::string &tag, int log_level) {
   this->log_levels_.push_back(LogLevelOverride{tag, log_level});
@@ -410,38 +164,12 @@ void Logger::add_on_log_callback(std::function<void(int, const char *, const cha
 }
 float Logger::get_setup_priority() const { return setup_priority::BUS + 500.0f; }
 const char *const LOG_LEVELS[] = {"NONE", "ERROR", "WARN", "INFO", "CONFIG", "DEBUG", "VERBOSE", "VERY_VERBOSE"};
-#ifdef USE_ESP32
-const char *const UART_SELECTIONS[] = {
-    "UART0",           "UART1",
-#if !defined(USE_ESP32_VARIANT_ESP32C3) && !defined(USE_ESP32_VARIANT_ESP32C6) && \
-    !defined(USE_ESP32_VARIANT_ESP32S2) && !defined(USE_ESP32_VARIANT_ESP32S3) && !defined(USE_ESP32_VARIANT_ESP32H2)
-    "UART2",
-#endif  // !USE_ESP32_VARIANT_ESP32C3 && !USE_ESP32_VARINT_ESP32C6 && !USE_ESP32_VARIANT_ESP32S2 &&
-        // !USE_ESP32_VARIANT_ESP32S3 && !USE_ESP32_VARIANT_ESP32H2
-#if defined(USE_ESP32_VARIANT_ESP32S2) || defined(USE_ESP32_VARIANT_ESP32S3)
-    "USB_CDC",
-#endif  // USE_ESP32_VARIANT_ESP32S2 || USE_ESP32_VARIANT_ESP32S3
-#if defined(USE_ESP32_VARIANT_ESP32C3) || defined(USE_ESP32_VARIANT_ESP32S3)
-    "USB_SERIAL_JTAG",
-#endif  // USE_ESP32_VARIANT_ESP32C3 || USE_ESP32_VARIANT_ESP32S3
-};
-#endif  // USE_ESP32
-#ifdef USE_ESP8266
-const char *const UART_SELECTIONS[] = {"UART0", "UART1", "UART0_SWAP"};
-#endif  // USE_ESP8266
-#ifdef USE_RP2040
-const char *const UART_SELECTIONS[] = {"UART0", "UART1", "USB_CDC"};
-#endif  // USE_RP2040
-#ifdef USE_LIBRETINY
-const char *const UART_SELECTIONS[] = {"DEFAULT", "UART0", "UART1", "UART2"};
-#endif  // USE_LIBRETINY
+
 void Logger::dump_config() {
   ESP_LOGCONFIG(TAG, "Logger:");
   ESP_LOGCONFIG(TAG, "  Level: %s", LOG_LEVELS[ESPHOME_LOG_LEVEL]);
   ESP_LOGCONFIG(TAG, "  Log Baud Rate: %" PRIu32, this->baud_rate_);
-#if defined(USE_ESP32) || defined(USE_ESP8266) || defined(USE_RP2040) || defined(USE_LIBRETINY)
-  ESP_LOGCONFIG(TAG, "  Hardware UART: %s", UART_SELECTIONS[this->uart_]);
-#endif
+  ESP_LOGCONFIG(TAG, "  Hardware UART: %s", get_uart_selection_());
 
   for (auto &it : this->log_levels_) {
     ESP_LOGCONFIG(TAG, "  Level for '%s': %s", it.tag.c_str(), LOG_LEVELS[it.level]);
